Add map section on bal page
