+++ pages/index.js
@@ -1,9 +1,7 @@
 import PropTypes from 'prop-types'
 import Link from 'next/link'
-import Image from 'next/image'
 
 import {getBALStats} from '@/lib/bal/api'
-import {numFormater} from '@/lib/format-numbers'
 
 import theme from '@/styles/theme'
 import Page from '@/layouts/main'
@@ -11,14 +9,12 @@ import Hero from '@/components/hero'
 import Section from '@/components/section'
 import SectionText from '@/components/section-text'
 import ButtonLink from '@/components/button-link'
-import Metric from '@/components/metric'
+import MapBalSection from '@/components/map-bal-section'
 import DocDownload from '@/components/doc-download'
 import Temoignages from '@/components/temoignages'
 import SocialMedia from '@/components/social-media'
 
 function Home({stats}) {
-  const populationCouvertePercent = Math.round((stats.bal.populationCouverte * 100) / stats.france.population)
-
   return (
     <Page>
       <Hero
@@ -74,48 +70,7 @@ function Home({stats}) {
       </Section>
 
       <Section background='color' title='État du déploiement des Bases Adresses Locales'>
-        <div className='deployement-container'>
-          <div className='map-illustration'>
-            <Link href='/deploiement-bal'>
-              <a>
-                <Image src='/images/france-illustration.svg' layout='responsive' height={400} width={400} />
-              </a>
-            </Link>
-          </div>
-          <div className='metrics-container'>
-            <Metric metric={stats.bal.nbCommunesCouvertes}> communes couvertes</Metric>
-            <Metric metric={numFormater(stats.bal.nbAdresses)}> d’adresses issues des BAL</Metric>
-            <Metric metric={populationCouvertePercent} isPercent>de la population couverte</Metric>
-          </div>
-          <ButtonLink href='/deploiement-bal' isOutlined color='white'>Carte de couverture des BAL</ButtonLink>
-        </div>
-        <style jsx>{`
-          .deployement-container {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            gap: 3em;
-            margin-top: 3em;
-          }
-
-          .map-illustration {
-            min-width: 290px;
-            max-width: 400px;
-          }
-
-          .metrics-container {
-            width: 100%;
-            display: grid;
-            justify-content: center;
-            grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-            padding: 1em;
-            gap: 3em;
-            text-align: center;
-            background: ${theme.colors.white};
-            color: ${theme.darkText};
-            border-radius: ${theme.borderRadius};
-          }
-        `}</style>
+        <MapBalSection stats={stats} />
       </Section>
 
       <Section background='grey' title='La fibre arrive dans la commune' subtitle='Communes et opérateurs, vous pouvez gagner du temps'>
@@ -172,6 +127,7 @@ Home.propTypes = {
     france: PropTypes.object.isRequired,
     bal: PropTypes.object.isRequired,
     ban: PropTypes.object.isRequired
-  }).isRequired}
+  }).isRequired
+}
 
 export default Home
